pyproject.toml
[project]
name = "overwatchlooker"
version = "0.1.0"
description = "Overwatch 2 screen analyzer using EasyOCR"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pynput>=1.8.1",
    "mss>=10.1.0",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "proc-tap>=1.0.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
    "anthropic>=0.83.0",
    "codex-open-client[pydantic]>=0.1.0",
    "litellm>=1.81.14",
    "telethon>=1.37.0",
    "questionary>=2.1.1",
    "pytesseract>=0.3.13",
    "dxcam>=0.3.0",
    "lz4>=4.4.5",
    "zstandard>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "screenshots: requires real screenshots in tests/fixtures",
    "live: calls real LLM APIs (opt-in, never runs by default)",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-timeout>=2.4.0",
]
